Fixed warnings and reformat (#17)
* BUILD_WITH_ALL_WARNINGS option to build with more warnings
* fixed most warnings while building with BUILD_WITH_ALL_WARNINGS
* reformatted most of the code
* fixed issues compiling with GCC 4.8
* INCLUDE_GENERATOR_RTCM=OFF now works

+++ generator/rtcm/encoder.cpp
@@ -32,7 +32,7 @@ void Encoder::u64(size_t bits, uint64_t value) {
     assert(bits > 0 && bits <= 64);
     for (size_t i = 0; i < bits; i++) {
         auto bit = (value >> (bits - i - 1)) & 1;
-        append_bit(bit);
+        append_bit(static_cast<uint8_t>(bit));
     }
 }
